+++ src/proto_008_PtEdo2Zk/lib_client_commands/client_proto_programs_commands.ml
@@ -425,10 +425,10 @@ let commands () =
     command
       ~group
       ~desc:"Ask the node to hash a Michelson script with `BLAKE2B`."
-      no_options
+      (args1 enforce_indentation_flag)
       (prefixes ["hash"; "script"] @@ file_or_literal_param () @@ stop)
-      (fun () expr_string (cctxt : Protocol_client_context.full) ->
-        let program = Michelson_v1_parser.parse_toplevel expr_string in
+      (fun check expr_string (cctxt : Protocol_client_context.full) ->
+        let program = Michelson_v1_parser.parse_toplevel ~check expr_string in
         Lwt.return @@ Micheline_parser.no_parsing_error program
         >>=? fun program ->
         let code = program.expanded in
@@ -733,18 +733,20 @@ let commands () =
       ~desc:
         "Conversion of Michelson script from Micheline, JSON or binary to \
          Micheline, JSON, binary or OCaml"
-      (args1 zero_loc_switch)
+      (args2 zero_loc_switch enforce_indentation_flag)
       (prefixes ["convert"; "script"]
       @@ file_or_literal_param () @@ prefix "from" @@ convert_input_format_param
       @@ prefix "to" @@ convert_output_format_param @@ stop)
-      (fun zero_loc
+      (fun (zero_loc, check)
            expr_string
            from_format
            to_format
            (cctxt : Protocol_client_context.full) ->
         (match from_format with
         | `Michelson ->
-            let program = Michelson_v1_parser.parse_toplevel expr_string in
+            let program =
+              Michelson_v1_parser.parse_toplevel ~check expr_string
+            in
             Lwt.return @@ Micheline_parser.no_parsing_error program
             >>=? fun program ->
             (typecheck_program