+++ src/proto_alpha/lib_client/client_proto_args.ml
@@ -504,6 +504,14 @@ let unparsing_mode_arg ~default =
     ~default
     unparsing_mode_parameter
 
+let enforce_indentation_flag =
+  switch
+    ~long:"enforce-indentation"
+    ~doc:
+      "Check that the Micheline expression passed to this command is \
+       well-indented."
+    ()
+
 module Daemon = struct
   let baking_switch =
     switch ~long:"baking" ~short:'B' ~doc:"run the baking daemon" ()